refactor: exception 패키지 변경 및 디폴트 메시지 추가
