+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,10 +1,11 @@
 package nextstep.reservation;
 
 import nextstep.member.Member;
-import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.*;
+import nextstep.support.excpetion.DuplicateReservationException;
+import nextstep.support.excpetion.NotExistReservationException;
+import nextstep.support.excpetion.NotQualifiedMemberException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
