+++ src/main/java/nextstep/support/LoginMemberArgumentResolver.java
@@ -2,16 +2,15 @@ package nextstep.support;
 
 import nextstep.auth.AuthorizationTokenExtractor;
 import nextstep.auth.JwtTokenProvider;
-import nextstep.member.Member;
 import nextstep.member.MemberDao;
+import nextstep.support.excpetion.InvalidAuthorizationTokenException;
+import nextstep.support.excpetion.NotExistMemberException;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-
-import java.util.Optional;
 
 @Component
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
