+++ src/main/java/nextstep/support/excpetion/DuplicateReservationException.java
@@ -1,8 +1,8 @@
-package nextstep.support;
+package nextstep.support.excpetion;
 
 public class DuplicateReservationException extends DuplicateEntityException {
     public DuplicateReservationException() {
-        super();
+        super("중복된 시간에 예약을 할 수 없습니다.");
     }
 
     public DuplicateReservationException(String message) {
